A new version more simple for C/S architechture
Separate the target run logic from the stack collection class.
Take advantage of the C++ feature definitions in the eBPF program header
file to simplify the code with macro.
Add a sub-function queue to control the use of multiple sub-functions at
the same time.
Renaming symbols makes the code easier to understand.
Removed the code that implemented the visualization capability.

Signed-off-by: LiuLingze <luiyanbing@foxmail.com>